+++ packages/advisor/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/advisor [1.0.0-alpha.127](https://github.com/anolilab/lunora/compare/@lunora/advisor@1.0.0-alpha.126...@lunora/advisor@1.0.0-alpha.127) (2026-09-12)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.120
+
 ## @lunora/advisor [1.0.0-alpha.126](https://github.com/anolilab/lunora/compare/@lunora/advisor@1.0.0-alpha.125...@lunora/advisor@1.0.0-alpha.126) (2026-09-11)