+++ packages/config/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/config [1.0.0-alpha.215](https://github.com/anolilab/lunora/compare/@lunora/config@1.0.0-alpha.214...@lunora/config@1.0.0-alpha.215) (2026-09-12)
+
+
+### Dependencies
+
+* **@lunora/codegen:** upgraded to 1.0.0-alpha.178
+* **@lunora/seed:** upgraded to 1.0.0-alpha.126
+* **@lunora/studio:** upgraded to 1.0.0-alpha.176
+
 ## @lunora/config [1.0.0-alpha.214](https://github.com/anolilab/lunora/compare/@lunora/config@1.0.0-alpha.213...@lunora/config@1.0.0-alpha.214) (2026-09-11)