+++ packages/payment/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/payment [1.0.0-alpha.123](https://github.com/anolilab/lunora/compare/@lunora/payment@1.0.0-alpha.122...@lunora/payment@1.0.0-alpha.123) (2026-09-12)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.120
+
 ## @lunora/payment [1.0.0-alpha.122](https://github.com/anolilab/lunora/compare/@lunora/payment@1.0.0-alpha.121...@lunora/payment@1.0.0-alpha.122) (2026-09-11)